# Add HasCallStack constraints to get informative errors and traces.