This improves the information in error messages, and would be necessary for tracing & analysis in jni.

Also has a commit to check for null pointers which could crash the program. Placed here to avoid conflicts but could be moved to another PR.