@facundominguez 2017-08-25
There is a penalty in performance.

Adding `HasCallStack` interferes with caching class references. I modified the benchmarks so they allocate the class references in a CAF and pass them to the `call` and `callStatic`.

With HasCallStack constraints:
```
benchmarking Calls/Java calls/static method call: unboxed single arg / unboxed return
time                 608.4 ns   (606.7 ns .. 610.6 ns)

benchmarking Calls/Java calls/jni static method call: unboxed single arg / unboxed return
time                 572.6 ns   (570.5 ns .. 575.1 ns)
                     1.000 R²   (1.000 R² .. 1.000 R²)

benchmarking Calls/Java calls/method call: no args / unboxed return
time                 6.003 μs   (5.876 μs .. 6.119 μs)

benchmarking Calls/Java calls/method call: boxed single arg / unboxed return
time                 13.89 μs   (13.15 μs .. 14.89 μs)
```

Without HasCallStack (i.e. without this PR), these are the numbers:
```
benchmarking Calls/Java calls/static method call: unboxed single arg / unboxed return
time                 521.9 ns   (519.9 ns .. 523.9 ns)
             
benchmarking Calls/Java calls/jni static method call: unboxed single arg / unboxed return
time                 499.8 ns   (498.2 ns .. 501.4 ns)
             
benchmarking Calls/Java calls/method call: no args / unboxed return
time                 1.711 μs   (1.701 μs .. 1.722 μs)
             
benchmarking Calls/Java calls/method call: boxed single arg / unboxed return
time                 7.159 μs   (6.367 μs .. 8.197 μs)
```