@facundominguez 2017-08-30
I suppose most of the cost should be explained by an increment of the calls to the GC, and I'll try to check it.

The alternatives I see:
1. Live with the performance penalty.
2. Ask GHC devs if it would be possible to disable `HasCallStack` with runtime flags.
3. Implement a library to collect stack traces with a mechanism that can be disabled at runtime.
4. Don't collect stack traces, and only collect the source location of call sites.