@mboes 2017-08-30
Insofar as we can live with (4), that would be ideal. Collecting call stacks is brittle to begin with, because it infects user code too, not just JNI bindings, `jvm` functions, inline-java, sparkle... etc etc. I believe that there are by now alternative ways to get stack traces when we do need them, in particular as a byproduct of the recent DWARF support effort.

Let's ping @bgamari and @scpmw about that.