@mboes 2017-08-30
> Insofar as we can live with (4), that would be ideal.

Well I should say... least bad, in my view.