@bgamari 2017-09-01
Right, since 8.2 you can reliably extract backtraces from running Haskell programs compiled with DWARF information (i.e. with GHC's `-g` flag). This functionality is exposed in Haskell through the `GHC.ExecutionStack` [module](https://hackage.haskell.org/package/base-4.10.0.0/docs/GHC-ExecutionStack.html). Note that there are a few caveats here,
 * GHC must be built with `libdw` support enabled (i.e. with the `--enable-dwarf-unwind` `configure` flag)
 * For reliable backtraces, all libraries linked against (including core libraries like `base`) must be built with `-g`
 * These are DWARF backtraces; the source spans will generally be more useful than the symbol names since the latter will often be compiled-generated unique names.