@mboes 2017-09-01
@bgamari many thanks for you input. A few followup questions:

* Is `libdw` support enabled by default?
* Does this work well even with pervasive use of FFI calls?
* Will DWARF backtraces contain only symbols names, and if so is there an easy way to map those back to source spans?