@bgamari 2017-09-01
> Is libdw support enabled by default?

We don't enable GHC's unwinding support by default since it requires that `libdw` be installed on the host.  I did, however, produce an [8.2.1 binary distribution](https://downloads.haskell.org/~ghc/8.2.1/ghc-8.2.1-x86_64-deb8-linux-dwarf.tar.xz) with it enabled (and all libraries built with `-g`).

> Does this work well even with pervasive use of FFI calls?

Yes, this was broken in 8.0 (which is why I advised against using it) but 8.2 should be solid.

> Will DWARF backtraces contain only symbols names, and if so is there an easy way to map those back to source spans?

The Haskell interface provides both symbol names and spans. See the haddocks for `GHC.ExecutionStack`.