@mboes 2017-09-01
Awesome!

Slightly off-topic question: @bgamari I gather from what you're saying that, given the quality of DWARF support now, on a host with libdw etc, in principle we could have a version of `base` where `error` provides good stack traces yet without needing `HasCallStack` constraints that pollute the whole call chain?