@bgamari 2017-09-01
Yep, this should indeed be possible. In the future it might also be possible to collect stack traces for arbitrary exceptions as well. I jotted down some ideas on this subject on the [Wiki](https://ghc.haskell.org/trac/ghc/wiki/Exceptions/StackTraces).