@facundominguez 2017-09-08
I tried the ghc binary with dwarf support. Doesn't seem to provide enough information yet.
```
$ cat t.hs
import GHC.ExecutionStack

f = getStackTrace

g = f

main = g >>= putStrLn . maybe "No stack trace" showST

showST = unlines . map (maybe "no loc info" showSrcLoc . srcLoc)

showSrcLoc (SrcLoc f l c) = concat [f, ":", show l, ":", show c]

$ ghc -g -rtsopts -fforce-recomp t.hs
[1 of 1] Compiling Main             ( t.hs, t.o )
Linking t ...

$ ./t
t: Failed to get stack frames of current process: no matching address range: Success
rts/Libdw.c:288:0
no loc info
no loc info
no loc info
no loc info
rts/Libdw.c:259:0
no loc info
```