@facundominguez 2017-09-11
Rebased master and added a patch with the changes I did to benchmark this.

The GC does have more work to do with `HasCallStack`.
```
$ stack --nix build jvm --bench --benchmark-arguments='-m glob "**/[sm]*ethod call*/*"'
...
  21,658,797,984 bytes allocated in the heap
     868,177,872 bytes copied during GC
       1,940,968 bytes maximum residency (460 sample(s))
         547,936 bytes maximum slop
               6 MB total memory in use (0 MB lost due to fragmentation)

                                     Tot time (elapsed)  Avg pause  Max pause
  Gen  0     28316 colls,     0 par    0.921s   0.918s     0.0000s    0.0005s
  Gen  1       460 colls,     0 par    0.468s   0.468s     0.0010s    0.0047s

  TASKS: 5 (1 bound, 4 peak workers (4 total), using -N1)

  SPARKS: 0 (0 converted, 0 overflowed, 0 dud, 0 GC'd, 0 fizzled)

  INIT    time    0.001s  (  0.001s elapsed)
  MUT     time   17.712s  ( 16.123s elapsed)
  GC      time    1.388s  (  1.386s elapsed)
  EXIT    time    0.001s  (  0.000s elapsed)
  Total   time   19.195s  ( 17.509s elapsed)

  Alloc rate    1,222,839,352 bytes per MUT second

  Productivity  92.8% of total user, 92.1% of total elapsed
```
without `HasCallStack`:
```
$ stack --nix build jvm --bench --benchmark-arguments='-m glob "**/[sm]*ethod call*/*"'
...
  12,803,902,856 bytes allocated in the heap
      34,427,952 bytes copied during GC
       2,066,576 bytes maximum residency (14 sample(s))
         621,040 bytes maximum slop
               6 MB total memory in use (0 MB lost due to fragmentation)

                                     Tot time (elapsed)  Avg pause  Max pause
  Gen  0     18123 colls,     0 par    0.133s   0.129s     0.0000s    0.0003s
  Gen  1        14 colls,     0 par    0.022s   0.022s     0.0016s    0.0037s

  TASKS: 5 (1 bound, 4 peak workers (4 total), using -N1)

  SPARKS: 0 (0 converted, 0 overflowed, 0 dud, 0 GC'd, 0 fizzled)

  INIT    time    0.001s  (  0.001s elapsed)
  MUT     time   17.682s  ( 17.429s elapsed)
  GC      time    0.155s  (  0.151s elapsed)
  EXIT    time    0.001s  (  0.001s elapsed)
  Total   time   17.969s  ( 17.582s elapsed)

  Alloc rate    724,114,250 bytes per MUT second

  Productivity  99.1% of total user, 99.1% of total elapsed
```